# docs/swagger.yaml
openapi: 3.0.3
info:
  title: User Data Processor API
  description: |
    API para processamento e gerenciamento de dados de usuários.
    
    Suporta:
    - Upload de arquivos em múltiplos formatos (CSV, JSON, XML)
    - Consulta de dados em múltiplos formatos (JSON, CSV, XML)
    - CRUD completo de usuários
    - Validação de dados

  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Servidor Local
  - url: http://localhost:8080/api
    description: API Base Path

tags:
  - name: Users
    description: Operações relacionadas a usuários
  - name: Health
    description: Verificação de saúde da aplicação

paths:
  /api/users/upload:
    post:
      tags:
        - Users
      summary: Upload de arquivo com dados de usuários
      description: |
        Faz upload de um arquivo contendo dados de usuários.
        
        Formatos aceitos:
        - CSV: arquivo com header (name, email)
        - JSON: array de objetos {name, email}
        - XML: <users><user><n>nome</n><email>email</email></user></users>

      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - fileType
              properties:
                file:
                  type: string
                  format: binary
                  description: Arquivo contendo dados de usuários
                fileType:
                  type: string
                  enum: [CSV, JSON, XML]
                  description: Tipo do arquivo enviado
            examples:
              csv:
                summary: Upload CSV
                value:
                  file: users.csv
                  fileType: CSV
              json:
                summary: Upload JSON
                value:
                  file: users.json
                  fileType: JSON
              xml:
                summary: Upload XML
                value:
                  file: users.xml
                  fileType: XML
      responses:
        '200':
          description: Arquivo processado com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              example:
                - id: 1
                  name: "John Doe"
                  email: "john.doe@example.com"
                  source: "CSV"
                  createdAt: "2025-11-05T18:30:00"
        '400':
          description: Erro de validação ou formato inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidFormat:
                  summary: Formato inválido
                  value:
                    timestamp: "2025-11-05T18:30:00"
                    status: 400
                    error: "Bad Request"
                    message: "Tipo de arquivo inválido"
                    path: "/api/users/upload"
                invalidEmail:
                  summary: Email inválido
                  value:
                    timestamp: "2025-11-05T18:30:00"
                    status: 400
                    error: "Bad Request"
                    message: "Email inválido"
                    path: "/api/users/upload"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users:
    get:
      tags:
        - Users
      summary: Listar todos os usuários
      description: Retorna lista de todos os usuários cadastrados
      operationId: getAllUsers
      parameters:
        - name: format
          in: query
          description: Formato de resposta desejado
          required: false
          schema:
            type: string
            enum: [json, csv, xml]
            default: json
      responses:
        '200':
          description: Lista de usuários retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              example:
                - id: 1
                  name: "John Doe"
                  email: "john.doe@example.com"
                  source: "CSV"
                  createdAt: "2025-11-05T18:30:00"
                - id: 2
                  name: "Jane Smith"
                  email: "jane.smith@example.com"
                  source: "JSON"
                  createdAt: "2025-11-05T18:31:00"
            text/csv:
              schema:
                type: string
              example: |
                createdAt,email,id,name,source
                "2025-11-05T18:30:00",john.doe@example.com,1,"John Doe",CSV
                "2025-11-05T18:31:00",jane.smith@example.com,2,"Jane Smith",JSON
            application/xml:
              schema:
                $ref: '#/components/schemas/UsersXmlResponse'
              example: |
                <users>
                  <user>
                    <id>1</id>
                    <name>John Doe</name>
                    <email>john.doe@example.com</email>
                    <source>CSV</source>
                    <createdAt>2025-11-05T18:30:00</createdAt>
                  </user>
                </users>
        '500':
          description: Erro ao formatar resposta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Buscar usuário por ID
      description: Retorna os dados de um usuário específico pelo seu ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: ID do usuário
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Usuário encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                id: 1
                name: "John Doe"
                email: "john.doe@example.com"
                source: "CSV"
                createdAt: "2025-11-05T18:30:00"
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                timestamp: "2025-11-05T18:30:00"
                status: 404
                error: "Not Found"
                message: "Usuário não encontrado com ID: 1"
                path: "/api/users/1"

    delete:
      tags:
        - Users
      summary: Deletar usuário
      description: Remove um usuário do sistema pelo seu ID
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: ID do usuário a ser deletado
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Usuário deletado com sucesso (sem conteúdo)
        '404':
          description: Usuário não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/users/source/{source}:
    get:
      tags:
        - Users
      summary: Buscar usuários por origem
      description: Retorna todos os usuários cadastrados a partir de uma determinada origem (CSV, JSON, XML)
      operationId: getUsersBySource
      parameters:
        - name: source
          in: path
          description: Origem dos dados (CSV, JSON, XML)
          required: true
          schema:
            type: string
            enum: [CSV, JSON, XML]
      responses:
        '200':
          description: Lista de usuários da origem especificada
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
              example:
                - id: 1
                  name: "John Doe"
                  email: "john.doe@example.com"
                  source: "CSV"
                  createdAt: "2025-11-05T18:30:00"
                - id: 2
                  name: "Jane Smith"
                  email: "jane.smith@example.com"
                  source: "CSV"
                  createdAt: "2025-11-05T18:31:00"
        '404':
          description: Nenhum usuário encontrado para a origem especificada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /actuator/health:
    get:
      tags:
        - Health
      summary: Verificar saúde da aplicação
      description: Endpoint de health check para monitoramento
      operationId: healthCheck
      responses:
        '200':
          description: Aplicação saudável
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "UP"
              example:
                status: "UP"

components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID único do usuário
          example: 1
        name:
          type: string
          description: Nome completo do usuário
          example: "John Doe"
        email:
          type: string
          format: email
          description: Email do usuário
          example: "john.doe@example.com"
        source:
          type: string
          enum: [CSV, JSON, XML]
          description: Origem dos dados do usuário
          example: "CSV"
        createdAt:
          type: string
          format: date-time
          description: Data e hora de criação do registro
          example: "2025-11-05T18:30:00"

    UserInput:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: Nome completo do usuário
          minLength: 1
          example: "John Doe"
        email:
          type: string
          format: email
          description: Email do usuário
          pattern: '^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$'
          example: "john.doe@example.com"

    UsersXmlResponse:
      type: object
      xml:
        name: users
      properties:
        user:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Momento em que o erro ocorreu
          example: "2025-11-05T18:30:00"
        status:
          type: integer
          description: Código HTTP do erro
          example: 400
        error:
          type: string
          description: Tipo do erro
          example: "Bad Request"
        message:
          type: string
          description: Mensagem descritiva do erro
          example: "Email inválido"
        path:
          type: string
          description: Caminho da requisição que gerou o erro
          example: "/api/users/upload"
        details:
          type: string
          nullable: true
          description: Detalhes adicionais do erro (opcional)
          example: null